+++ packages/photo-companion/src/lib/components/bridges-info/bridge-state-list.tsx
@@ -3,8 +3,8 @@ import { createMemo, For, Show } from "solid-js";
 import { objectEntries } from "utils";
 
 import { CardEntry } from "~/lib/components";
-import { useTranslation } from "~/lib/context/translation";
 import { useDatetime } from "~/lib/hooks";
+import { useTranslation } from "~/services/translation";
 
 import { BridgesStateAll } from "./bridge-state-all";
