+++ packages/photo-companion/src/lib/components/bridges-info/next-bridge-countdown.tsx
@@ -2,8 +2,8 @@ import { getNextBridgeEvent } from "bridge-schedule";
 import { createMemo} from "solid-js";
 
 import { CardEntry } from "~/lib/components";
-import { useTranslation } from "~/lib/context/translation";
 import { createCountdown, useDatetime } from "~/lib/hooks";
+import { useTranslation } from "~/services/translation";
 
 /**
  * Renders countdown till the next bridge event.
