+++ packages/photo-companion/src/lib/components/timeline-event/timeline-event-bridge.ts
@@ -2,8 +2,8 @@ import type { BridgeEvent } from "types";
 
 import { Bridge } from "../bridge/bridge";
 
-import type { Translation } from "~/lib/context/translation";
 import type { EventComponent } from "~/lib/types";
+import type { Translation } from "~/services/translation";
 
 export function bridgeEventComponent(event: BridgeEvent, t: Translation): EventComponent<{ open: boolean }> {
 	return {
