+++ packages/photo-companion/src/lib/components/timeline-event/timeline-event-empty.tsx
@@ -1,6 +1,6 @@
 import { classnames } from "utils";
 
-import { useTranslation } from "~/lib/context/translation";
+import { useTranslation } from "~/services/translation";
 
 import styles from "./event.module.css";
