+++ packages/photo-companion/src/lib/components/timeline-event/timeline-event-moon.ts
@@ -4,8 +4,8 @@ import type { MoonEvent } from "types";
 
 import { Moon } from "../moon/moon";
 
-import type { Translation } from "~/lib/context/translation";
 import type { EventComponent } from "~/lib/types";
+import type { Translation } from "~/services/translation";
 
 export function moonEventComponent(event: MoonEvent, t: Translation): EventComponent<{ phase: number, rotation: number }> {
 	const { azimuth, fraction, phase, rotation, waxing } = event.data;
