+++ packages/photo-companion/src/lib/components/toast-suggest-update/toast-suggest-update.tsx
@@ -2,8 +2,8 @@ import { createEffect, on } from "solid-js";
 import { toast } from "solid-sonner";
 import { Button } from "ui";
 
-import { useTranslation } from "~/lib/context/translation";
 import { useServiceWorker } from "~/services/service-worker";
+import { useTranslation } from "~/services/translation";
 
 import styles from "./toast-suggest-update.module.css";
