+++ packages/photo-companion/src/routes/about/about.page.tsx
@@ -1,7 +1,7 @@
 import { lazy, Match, Switch } from "solid-js";
 
 import { Markdown } from "~/lib/components/markdown";
-import { useTranslation } from "~/lib/context/translation";
+import { useTranslation } from "~/services/translation";
 
 const PageAboutEn = lazy(() => import("./about.en.md"));
 const PageAboutRu = lazy(() => import("./about.ru.md"));
