+++ packages/photo-companion/src/routes/bridges/components/card-bridge/card-bridge-sparkline.tsx
@@ -1,9 +1,9 @@
 import { For, mergeProps } from "solid-js";
 import { round } from "utils/math";
 
-import { useTranslation } from "~/lib/context/translation";
 import { getDate } from "~/lib/helpers";
 import { createWavyPath } from "~/lib/helpers/svg";
+import { useTranslation } from "~/services/translation";
 
 import { DEFAULT_PROPS_SPARKLINE } from "./card-bridge.const";
 import { buildBridgeSparklinePath } from "./card-bridge.utils";
