+++ packages/photo-companion/src/routes/bridges/components/card-bridge/card-bridge.tsx
@@ -4,8 +4,8 @@ import { IconWarning } from "ui";
 
 import type { BridgeName, BridgeState } from "types";
 
-import { useTranslation } from "~/lib/context/translation";
 import { createCountdown, useDatetime } from "~/lib/hooks";
+import { useTranslation } from "~/services/translation";
 
 import { BridgeSparkline } from "./card-bridge-sparkline";
