+++ packages/photo-companion/src/routes/height-by-shadow/height-by-shadow.page.tsx
@@ -1,6 +1,6 @@
 import { IconShadow } from "ui";
 
-import { useTranslation } from "~/lib/context/translation";
+import { useTranslation } from "~/services/translation";
 
 import { Error } from "./components/error/error";
 import { Form } from "./components/form/form";
