+++ packages/photo-companion/src/routes/moon/moon-altitude.tsx
@@ -2,7 +2,7 @@
 import { getMoonPosition, getSunPosition } from "moon-sun-calc";
 
 import { ElevationGraph } from "~/lib/components";
-import { useTranslation } from "~/lib/context/translation";
+import { useTranslation } from "~/services/translation";
 
 import styles from "./moon.module.css";
