+++ packages/photo-companion/src/routes/moon/moon-times.tsx
@@ -1,7 +1,7 @@
 import { classnames } from "utils";
 
 import { GaugeTime, Moon } from "~/lib/components";
-import { useTranslation } from "~/lib/context/translation";
+import { useTranslation } from "~/services/translation";
 
 import type { MoonData } from "./use-moon-data";
