+++ packages/photo-companion/src/routes/now/components/now-moon.tsx
@@ -1,7 +1,7 @@
 import { Show } from "solid-js";
 
 import { CardEntry, CardInfo } from "~/lib/components";
-import { useTranslation } from "~/lib/context/translation";
+import { useTranslation } from "~/services/translation";
 
 import { useMoonData } from "../../moon/use-moon-data";
