+++ packages/photo-companion/src/routes/settings/settings.form.tsx
@@ -1,7 +1,7 @@
 import { Button, Fieldset, Form, InputRadio } from "ui";
 
-import { useTranslation } from "~/lib/context/translation";
 import { SETTINGS_CITY_PRESETS } from "~/services/settings";
+import { useTranslation } from "~/services/translation";
 
 import { InputEventFilters, InputGeolocation, InputTabsSelect } from "./components";
 import { CITY_OPTIONS, FORM_INPUT_NAME, LANGUAGE_OPTIONS } from "./settings.const";
