+++ packages/photo-companion/src/routes/sun/sun-altitude.tsx
@@ -2,7 +2,7 @@
 import { getSunPosition } from "moon-sun-calc";
 
 import { ElevationGraph } from "~/lib/components";
-import { useTranslation } from "~/lib/context/translation";
+import { useTranslation } from "~/services/translation";
 
 interface SunAltitudeProps {
 	date: Date;
