+++ packages/photo-companion/src/routes/sun/use-sun-data.ts
@@ -3,9 +3,9 @@ import { createMemo } from "solid-js";
 import { calcDuration } from "utils/date";
 import { round } from "utils/math";
 
-import { useTranslation } from "~/lib/context/translation";
 import { useDatetime } from "~/lib/hooks";
 import { useSettings } from "~/services/settings";
+import { useTranslation } from "~/services/translation";
 
 export interface SunData {
 	blueHourDawn: string;
