+++ packages/photo-companion/src/services/moon/index.ts
@@ -4,7 +4,7 @@ import { isNullable } from "utils/validators";
 
 import type { MoonEvent } from "types";
 
-import { useTranslation } from "~/lib/context/translation";
+import { useTranslation } from "~/services/translation";
 
 export const getMoonEvents = (date: Date = new Date(), latitude: number, longitude: number): MoonEvent[] => {
 	const events: MoonEvent[] = [];
