+++ packages/photo-companion/src/services/translation/translation.en.ts
