+++ packages/photo-companion/src/services/translation/translation.type.ts
