+++ packages/photo-companion/src/services/translation/translation.types.ts
